requirements.txt
Flask==2.1.0
Flask-RESTful==0.3.9
redis
flask-cors
flask_restful
mysql-connector-python
